Move test stuub classes into their own files
This brings us into PSR12 compliance
