+++ tests/Checks/DatabaseHealthCheckTest.php
@@ -2,12 +2,10 @@
 
 namespace Tests\Checks;
 
-use Exception;
-use Illuminate\Database\Connection;
-use Illuminate\Database\ConnectionResolver;
-use Illuminate\Database\DatabaseManager as IlluminateDatabaseManager;
 use Illuminate\Foundation\Application;
-use InvalidArgumentException;
+use Tests\Stubs\Database\BadConnection;
+use Tests\Stubs\Database\DatabaseManager;
+use Tests\Stubs\Database\HealthyConnection;
 use Tests\TestCase;
 use UKFast\HealthCheck\Checks\DatabaseHealthCheck;
 use UKFast\HealthCheck\HealthCheckServiceProvider;
@@ -65,62 +63,5 @@ class DatabaseHealthCheckTest extends TestCase
 
         $this->assertTrue($status->isProblem());
         $this->assertSame('bad', $status->context()['connection']);
-    }
-}
-
-class HealthyConnection extends Connection
-{
-    public function __construct()
-    {
-    }
-
-    public function getPdo(): bool
-    {
-        return true;
-    }
-}
-
-class BadConnection extends Connection
-{
-    public function __construct()
-    {
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function getPdo(): never
-    {
-        throw new Exception();
-    }
-}
-
-class DatabaseManager extends IlluminateDatabaseManager
-{
-    protected $connections = [];
-
-    public function __construct()
-    {
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    public function connection($name = null)
-    {
-        if (!$name) {
-            return $this->connection('default');
-        }
-
-        if (!isset($this->connections[$name])) {
-            throw new InvalidArgumentException("Database [$name] not configured.");
-        }
-
-        return $this->connections[$name];
-    }
-
-    public function addConnection($name, $connection)
-    {
-        $this->connections[$name] = $connection;
     }
 }
