+++ tests/Checks/StorageHealthCheckTest.php
@@ -2,8 +2,8 @@
 
 namespace Tests\Checks;
 
-use Exception;
 use Illuminate\Foundation\Application;
+use Tests\Stubs\Storage\BadDisk;
 use Tests\TestCase;
 use Illuminate\Support\Facades\Storage;
 use UKFast\HealthCheck\Checks\StorageHealthCheck;
@@ -64,16 +64,5 @@ class StorageHealthCheckTest extends TestCase
         $status = (new StorageHealthCheck($this->app))->status();
 
         $this->assertTrue($status->isOkay());
-    }
-}
-
-class BadDisk
-{
-    /**
-     * @throws Exception
-     */
-    public function __call($name, $arguments): never
-    {
-        throw new Exception();
     }
 }
