+++ tests/Stubs/Cache/BadStore.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Tests\Stubs\Cache;
+
+use Exception;
+
+class BadStore
+{
+    /**
+     * @throws Exception
+     */
+    public function __call($name, $arguments): never
+    {
+        throw new Exception();
+    }
+}
