+++ tests/Stubs/Database/DatabaseManager.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Tests\Stubs\Database;
+
+use Illuminate\Database\DatabaseManager as IlluminateDatabaseManager;
+use InvalidArgumentException;
+
+class DatabaseManager extends IlluminateDatabaseManager
+{
+    protected $connections = [];
+
+    public function __construct()
+    {
+    }
+
+    /**
+     * @throws InvalidArgumentException
+     */
+    public function connection($name = null)
+    {
+        if (!$name) {
+            return $this->connection('default');
+        }
+
+        if (!isset($this->connections[$name])) {
+            throw new InvalidArgumentException("Database [$name] not configured.");
+        }
+
+        return $this->connections[$name];
+    }
+
+    public function addConnection($name, $connection)
+    {
+        $this->connections[$name] = $connection;
+    }
+}
